ctr: move resolver to commands package
